+++ modules/catalog/factory/go.mod
@@ -1,0 +1,3 @@
+module github.com/treeverse/lakefs/modules/catalog/factory
+
+go 1.24.4
