--- modules/graveler/factory/go.mod
@@ -1,3 +1,0 @@
-module github.com/treeverse/lakefs/modules/graveler/factory
-
-go 1.24.4
